Move buildFileTree into a fileTree reducer
