+++ src/components/FileTree/index.tsx
@@ -11,158 +11,68 @@ import FileTreeNode, {
 import Loading from '../Loading';
 import { ApplicationState, ConnectedReduxProps } from '../../configureStore';
 import {
+  DirectoryNode,
+  TreeNode,
+  actions as fileTreeActions,
+  getTree,
+} from '../../reducers/fileTree';
+import {
   Version,
   actions as versionsActions,
   getVersionInfo,
 } from '../../reducers/versions';
-import { getLocalizedString, gettext } from '../../utils';
+import { gettext } from '../../utils';
 
-type FileNode = {
-  id: string;
-  name: string;
-};
-
-export type DirectoryNode = {
-  id: string;
-  name: string;
-  children: TreeNode[];
-};
-
-export type TreeNode = FileNode | DirectoryNode;
+type LoadData = () => void;
 
 export type TreefoldRenderPropsForFileTree = TreefoldRenderProps<TreeNode>;
 
-const recursiveSortInPlace = (node: DirectoryNode): void => {
-  node.children.sort((a, b) => {
-    if ('children' in a && 'children' in b) {
-      return a.name.localeCompare(b.name);
-    }
-
-    if ('children' in a === false && 'children' in b === false) {
-      return a.name.localeCompare(b.name);
-    }
-
-    if ('children' in b === false) {
-      return -1;
-    }
-
-    return 1;
-  });
-
-  node.children.forEach((child) => {
-    if ('children' in child) {
-      recursiveSortInPlace(child);
-    }
-  });
-};
-
-const getVersionName = (version: Version) => {
-  return getLocalizedString(version.addon.name);
-};
-
-export const getRootPath = (version: Version) => {
-  return `root-${getVersionName(version)}`;
-};
-
-export const buildFileTree = (version: Version): DirectoryNode => {
-  const { entries } = version;
-  const root: DirectoryNode = {
-    id: getRootPath(version),
-    name: getVersionName(version),
-    children: [],
-  };
-
-  // We need to know how depth the tree is because we'll build the Treebeard
-  // tree depth by depth.
-  const maxDepth = entries.reduce((max, entry) => {
-    if (entry.depth > max) {
-      return entry.depth;
-    }
-
-    return max;
-  }, 0);
-
-  let currentDepth = 0;
-  while (currentDepth <= maxDepth) {
-    // We find the entries for the current depth.
-    const currentEntries = entries.filter(
-      // eslint-disable-next-line no-loop-func
-      (entry) => entry.depth === currentDepth,
-    );
-
-    // This is where we create new "nodes" for each entry.
-    // eslint-disable-next-line no-loop-func
-    currentEntries.forEach((entry) => {
-      let currentNode = root;
-
-      if (currentDepth > 0) {
-        // We need to find the current node (directory) to add the current
-        // entry in its children. We do this by splitting the `path` attribute
-        // and visit each node until we reach the desired node.
-        //
-        // This only applies when the current depth is not 0 (a.k.a. the root
-        // directory) because we already know `root`.
-        const parts = entry.path.split('/');
-        // Remove the filename
-        parts.pop();
-
-        for (let i = 0; i < parts.length; i++) {
-          const foundNode = currentNode.children.find(
-            (child: TreeNode) => child.name === parts[i],
-          ) as DirectoryNode;
-
-          if (foundNode) {
-            currentNode = foundNode;
-          }
-
-          // TODO: this should not happen but what if we don't find a node?
-        }
-      }
-
-      // Create a new node.
-      let node: TreeNode = {
-        id: entry.path,
-        name: entry.filename,
-      };
-
-      // When the entry is a directory, we create a `DirectoryNode`.
-      if (entry.type === 'directory') {
-        node = {
-          ...node,
-          children: [],
-        };
-      }
-
-      currentNode.children.push(node);
-    });
-
-    // To the next depth.
-    currentDepth++;
-  }
-
-  recursiveSortInPlace(root);
-
-  return root;
-};
-
 export type PublicProps = {
+  _loadData?: LoadData;
   onSelect: FileTreeNodeProps['onSelect'];
   versionId: number;
 };
 
 export type DefaultProps = {
+  _buildTree: typeof fileTreeActions.buildTree;
   _log: typeof log;
 };
 
 type PropsFromState = {
+  tree: DirectoryNode | void;
   version: Version | void;
 };
 
 type Props = PublicProps & DefaultProps & PropsFromState & ConnectedReduxProps;
 
 export class FileTreeBase extends React.Component<Props> {
+  loadData: LoadData;
+
   static defaultProps: DefaultProps = {
+    _buildTree: fileTreeActions.buildTree,
     _log: log,
+  };
+
+  constructor(props: Props) {
+    super(props);
+    // Allow dependency injection to test all the ways loadData() gets executed.
+    this.loadData = props._loadData || this._loadData;
+  }
+
+  componentDidMount() {
+    this.loadData();
+  }
+
+  componentDidUpdate() {
+    this.loadData();
+  }
+
+  _loadData = () => {
+    const { _buildTree, dispatch, tree, version } = this.props;
+
+    if (version && !tree) {
+      dispatch(_buildTree({ version }));
+    }
   };
 
   renderNode = (props: TreefoldRenderPropsForFileTree) => {
@@ -224,13 +134,11 @@ export class FileTreeBase extends React.Component<Props> {
   };
 
   render() {
-    const { version } = this.props;
+    const { tree, version } = this.props;
 
-    if (!version) {
+    if (!version || !tree) {
       return <Loading message={gettext('Loading version...')} />;
     }
-
-    const tree = buildFileTree(version);
 
     return (
       <ListGroup>
@@ -283,6 +191,7 @@ const mapStateToProps = (
   }
 
   return {
+    tree: getTree(state.fileTree, version),
     version,
   };
 };
